Move zombie session.(cc|hh) to webserver.hh
