+++ pdns/communicator.cc
@@ -32,7 +32,6 @@
 #include "logger.hh"
 #include "dns.hh"
 #include "arguments.hh"
-#include "session.hh"
 #include "packetcache.hh"
 #include <boost/lexical_cast.hpp>
